requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
django==5.2.6
djangorestframework==3.16.1
fonttools==4.59.2
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
response==0.5.0
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
